+++ internal/config/local.go
@@ -11,9 +11,11 @@ type (
 		TlsConfig           TlsConfig           `yaml:"tlsConfig"`
 		WiredConfig         EthernetConfig      `yaml:"wiredConfig"`
 		WifiConfigs         []WifiConfig        `yaml:"wifiConfigs"`
+		WiFiSyncEnabled     bool                `yaml:"wifiSyncEnabled"`
 		Ieee8021xConfigs    []Ieee8021xConfig   `yaml:"ieee8021xConfigs"`
 		ACMSettings         ACMSettings         `yaml:"acmactivate"`
 		EnterpriseAssistant EnterpriseAssistant `yaml:"enterpriseAssistant"`
+		CCMSettings         CCMSettings         `yaml:"ccmactivate"`
 	}
 	TlsConfig struct {
 		Delay int    `yaml:"delay" env-default:"3"`
@@ -68,5 +70,8 @@ type (
 		EAUsername   string `yaml:"eaUsername"`
 		EAPassword   string `yaml:"eaPassword"`
 		EAConfigured bool   `yaml:"eaConfigured"`
+	}
+	CCMSettings struct {
+		AMTPassword string `yaml:"amtPassword"`
 	}
 )