+++ internal/flags/flags.go
@@ -7,6 +7,7 @@ package flags
 import (
 	"bytes"
 	"encoding/base64"
+	"encoding/json"
 	"flag"
 	"fmt"
 	"net"
@@ -19,6 +20,9 @@ import (
 	"strconv"
 	"strings"
 	"time"
+
+	configv2 "github.com/open-amt-cloud-toolkit/go-wsman-messages/v2/pkg/config"
+	"github.com/open-amt-cloud-toolkit/go-wsman-messages/v2/pkg/security"
 
 	"github.com/google/uuid"
 	"github.com/ilyakaznacheev/cleanenv"
@@ -76,8 +80,11 @@ type Flags struct {
 	UseACM                              bool
 	EchoPass                            bool
 	configContent                       string
+	configContentV2                     string
+	configV2Key                         string
 	UUID                                string
 	LocalConfig                         config.Config
+	LocalConfigV2                       configv2.Configuration
 	amtInfoCommand                      *flag.FlagSet
 	amtActivateCommand                  *flag.FlagSet
 	amtDeactivateCommand                *flag.FlagSet
@@ -92,7 +99,7 @@ type Flags struct {
 	flagSetEnableWifiPort               *flag.FlagSet
 	flagSetMEBx                         *flag.FlagSet
 	flagSetAMTFeatures                  *flag.FlagSet
-	amtCommand                          amt.AMTCommand
+	AmtCommand                          amt.AMTCommand
 	netEnumerator                       NetEnumerator
 	IpConfiguration                     IPConfiguration
 	HostnameInfo                        HostnameInfo
@@ -108,6 +115,8 @@ type Flags struct {
 	KVM                                 bool
 	SOL                                 bool
 	IDER                                bool
+	LocalTlsEnforced                    bool
+	ControlMode                         int
 }
 
 func NewFlags(args []string, pr utils.PasswordReader) *Flags {
@@ -135,7 +144,7 @@ func NewFlags(args []string, pr utils.PasswordReader) *Flags {
 	flags.flagSetMEBx = flag.NewFlagSet(utils.SubCommandSetMEBx, flag.ContinueOnError)
 	flags.flagSetAMTFeatures = flag.NewFlagSet(utils.SubCommandSetAMTFeatures, flag.ContinueOnError)
 
-	flags.amtCommand = amt.NewAMTCommand()
+	flags.AmtCommand = amt.NewAMTCommand()
 	flags.netEnumerator = NetEnumerator{}
 	flags.netEnumerator.Interfaces = net.Interfaces
 	flags.netEnumerator.InterfaceAddrs = (*net.Interface).Addrs
@@ -343,5 +352,27 @@ func (f *Flags) handleLocalConfig() error {
 			return err
 		}
 	}
+	return nil
+}
+
+func (f *Flags) handleLocalConfigV2() error {
+	if f.configV2Key == "" {
+		log.Error("config error: missing encryption key")
+		return utils.FailedReadingConfiguration
+	}
+
+	security := security.Crypto{EncryptionKey: f.configV2Key}
+	content, err := security.ReadAndDecryptFile(f.configContentV2)
+	if err != nil {
+		log.Error("config error: ", err)
+		return err
+	}
+
+	_, err = json.MarshalIndent(content, "", "  ")
+	if err != nil {
+		log.Error("error formatting config content: ", err)
+		return err
+	}
+	f.LocalConfigV2 = content
 	return nil
 }