+++ internal/local/amt/wsman.go
@@ -6,6 +6,7 @@
 package amt
 
 import (
+	cryptotls "crypto/tls"
 	"encoding/base64"
 	"net"
 	"rpc/pkg/utils"
@@ -34,7 +35,7 @@ import (
 )
 
 type WSMANer interface {
-	SetupWsmanClient(username string, password string, logAMTMessages bool)
+	SetupWsmanClient(username string, password string, useTLS bool, logAMTMessages bool, tlsConfig *cryptotls.Config) error
 	Unprovision(int) (setupandconfiguration.Response, error)
 	GetGeneralSettings() (general.Response, error)
 	HostBasedSetupService(digestRealm string, password string) (hostbasedsetup.Response, error)
@@ -59,7 +60,7 @@ type WSMANer interface {
 	// WiFi
 	GetWiFiSettings() ([]wifi.WiFiEndpointSettingsResponse, error)
 	DeleteWiFiSetting(instanceId string) error
-	EnableWiFi() error
+	EnableWiFi(enableSync bool) error
 	AddWiFiSettings(wifiEndpointSettings wifi.WiFiEndpointSettingsRequest, ieee8021xSettings models.IEEE8021xSettings, wifiEndpoint, clientCredential, caCredential string) (wifiportconfiguration.Response, error)
 	// Wired
 	GetEthernetSettings() ([]ethernetport.SettingsResponse, error)
@@ -96,29 +97,43 @@ func NewGoWSMANMessages(lmsAddress string) *GoWSMANMessages {
 	}
 }
 
-func (g *GoWSMANMessages) SetupWsmanClient(username string, password string, logAMTMessages bool) {
+func (g *GoWSMANMessages) SetupWsmanClient(username string, password string, useTLS bool, logAMTMessages bool, tlsConfig *cryptotls.Config) error {
 	clientParams := client.Parameters{
 		Target:         g.target,
 		Username:       username,
 		Password:       password,
 		UseDigest:      true,
-		UseTLS:         false,
+		UseTLS:         useTLS,
+		TlsConfig:      tlsConfig,
 		LogAMTMessages: logAMTMessages,
 	}
-	logrus.Info("Attempting to connect to LMS...")
-	port := utils.LMSPort
+
+	var (
+		con net.Conn // Declared to hold the connection object
+		err error
+	)
 	if clientParams.UseTLS {
-		port = client.TLSPort
+		logrus.Info("Attempting to connect to LMS over TLS...")
+		con, err = cryptotls.Dial("tcp", utils.LMSAddress+":"+client.TLSPort, clientParams.TlsConfig)
+	} else {
+		logrus.Info("Attempting to connect to LMS...")
+		con, err = net.Dial("tcp", utils.LMSAddress+":"+client.NonTLSPort)
 	}
-	con, err := net.Dial("tcp4", utils.LMSAddress+":"+port)
+
 	if err != nil {
-		logrus.Info("Failed to connect to LMS, using local transport instead.")
+		logrus.Trace("Failed to connect to LMS: ", err)
+		if clientParams.UseTLS {
+			return utils.LMSConnectionFailed
+		}
+		logrus.Info("using local transport instead...")
 		clientParams.Transport = NewLocalTransport()
 	} else {
 		logrus.Info("Successfully connected to LMS.")
 		con.Close()
 	}
+
 	g.wsmanMessages = wsman.NewMessages(clientParams)
+	return nil
 }
 
 func (g *GoWSMANMessages) GetGeneralSettings() (general.Response, error) {
@@ -289,14 +304,20 @@ func (g *GoWSMANMessages) DeleteKeyPair(instanceID string) error {
 	_, err := g.wsmanMessages.AMT.PublicKeyManagementService.Delete(instanceID)
 	return err
 }
-func (g *GoWSMANMessages) EnableWiFi() error {
+func (g *GoWSMANMessages) EnableWiFi(enableSync bool) error {
 	response, err := g.wsmanMessages.AMT.WiFiPortConfigurationService.Get()
 	if err != nil {
 		return err
 	}
 
+	// Determine the sync state based on input parameter
+	syncState := wifiportconfiguration.LocalSyncDisabled
+	if enableSync {
+		syncState = wifiportconfiguration.UnrestrictedSync
+	}
+
 	// if local sync not enable, enable it
-	if response.Body.WiFiPortConfigurationService.LocalProfileSynchronizationEnabled == wifiportconfiguration.LocalSyncDisabled {
+	if response.Body.WiFiPortConfigurationService.LocalProfileSynchronizationEnabled != syncState {
 		putRequest := wifiportconfiguration.WiFiPortConfigurationServiceRequest{
 			RequestedState:                     response.Body.WiFiPortConfigurationService.RequestedState,
 			EnabledState:                       response.Body.WiFiPortConfigurationService.EnabledState,
@@ -306,7 +327,7 @@ func (g *GoWSMANMessages) EnableWiFi() error {
 			SystemName:                         response.Body.WiFiPortConfigurationService.SystemName,
 			CreationClassName:                  response.Body.WiFiPortConfigurationService.CreationClassName,
 			Name:                               response.Body.WiFiPortConfigurationService.Name,
-			LocalProfileSynchronizationEnabled: wifiportconfiguration.UnrestrictedSync,
+			LocalProfileSynchronizationEnabled: syncState,
 			LastConnectedSsidUnderMeControl:    response.Body.WiFiPortConfigurationService.LastConnectedSsidUnderMeControl,
 			NoHostCsmeSoftwarePolicy:           response.Body.WiFiPortConfigurationService.NoHostCsmeSoftwarePolicy,
 			UEFIWiFiProfileShareEnabled:        response.Body.WiFiPortConfigurationService.UEFIWiFiProfileShareEnabled,