+++ internal/local/info.go
@@ -7,12 +7,14 @@ package local
 
 import (
 	"bytes"
+	"crypto/tls"
 	"encoding/json"
 	"errors"
 	"fmt"
 	"net"
 	"os"
 	"rpc/internal/amt"
+	"rpc/internal/config"
 	"rpc/internal/flags"
 	"rpc/pkg/utils"
 	"strconv"
@@ -20,7 +22,6 @@ import (
 
 	"github.com/open-amt-cloud-toolkit/go-wsman-messages/v2/pkg/wsman/amt/publickey"
 	"github.com/open-amt-cloud-toolkit/go-wsman-messages/v2/pkg/wsman/amt/publicprivate"
-	"github.com/sirupsen/logrus"
 	log "github.com/sirupsen/logrus"
 )
 
@@ -162,18 +163,28 @@ func (service *ProvisioningService) DisplayAMTInfo() (err error) {
 		service.PrintOutput("Control Mode		: " + string(utils.InterpretControlMode(result)))
 	}
 	if service.flags.AmtInfo.OpState {
-		result, err := cmd.GetChangeEnabled()
+		majorVersion, err := GetMajorVersion(dataStruct["amt"].(string))
 		if err != nil {
 			log.Error(err)
 		}
-		if result.IsNewInterfaceVersion() {
-			opStateValue := "disabled"
-			if result.IsAMTEnabled() {
-				opStateValue = "enabled"
+		const minimumAMTVersion = 11
+		// Check if the AMT major version is greater than 11
+		if majorVersion > minimumAMTVersion {
+			result, err := cmd.GetChangeEnabled()
+			if err != nil {
+				log.Error(err)
 			}
+			if result.IsNewInterfaceVersion() {
+				opStateValue := "disabled"
+				if result.IsAMTEnabled() {
+					opStateValue = "enabled"
+				}
 
-			dataStruct["operationalState"] = opStateValue
-			service.PrintOutput("Operational State	: " + opStateValue)
+				dataStruct["operationalState"] = opStateValue
+				service.PrintOutput("Operational State	: " + opStateValue)
+			}
+		} else {
+			log.Debug("OpState will not work on AMT versions 11 and below.")
 		}
 	}
 	if service.flags.AmtInfo.DNS {
@@ -300,7 +311,11 @@ func (service *ProvisioningService) DisplayAMTInfo() (err error) {
 		}
 	}
 	if service.flags.AmtInfo.UserCert {
-		service.interfacedWsmanMessage.SetupWsmanClient("admin", service.flags.Password, logrus.GetLevel() == logrus.TraceLevel)
+		tlsConfig := &tls.Config{}
+		if service.flags.LocalTlsEnforced {
+			tlsConfig = config.GetTLSConfig(&service.flags.ControlMode)
+		}
+		service.interfacedWsmanMessage.SetupWsmanClient("admin", service.flags.Password, service.flags.LocalTlsEnforced, log.GetLevel() == log.TraceLevel, tlsConfig)
 		userCerts, _ := service.interfacedWsmanMessage.GetPublicKeyCerts()
 		userCertMap := map[string]publickey.RefinedPublicKeyCertificateResponse{}
 		for i := range userCerts {
@@ -418,4 +433,17 @@ func DecodeAMT(version, SKU string) string {
 		}
 	}
 	return result
+}
+func GetMajorVersion(version string) (int, error) {
+	amtParts := strings.Split(version, ".")
+	if len(amtParts) <= 1 {
+		return 0, fmt.Errorf("invalid AMT version format")
+	}
+
+	majorVersion, err := strconv.Atoi(amtParts[0])
+	if err != nil {
+		return 0, fmt.Errorf("invalid AMT version")
+	}
+
+	return majorVersion, nil
 }