+++ pkg/utils/constants.go
@@ -19,7 +19,8 @@ const (
 	// LMSAddress is used for determing what address to connect to LMS on
 	LMSAddress = "localhost"
 	// LMSPort is used for determining what port to connect to LMS on
-	LMSPort = "16992"
+	LMSPort    = "16992"
+	LMSTLSPort = "16993"
 
 	AMTUserName = "admin"
 
@@ -54,7 +55,7 @@ const (
 
 // (1-99) General Errors
 
-// (1-19) Basic errors outside of Open AMT Cloud Toolkit
+// (1-19) Basic errors outside of Device Management Toolkit
 var IncorrectPermissions = CustomError{Code: 1, Message: "IncorrectPermissions"}
 var HECIDriverNotDetected = CustomError{Code: 2, Message: "HECIDriverNotDetected"}
 var AmtNotDetected = CustomError{Code: 3, Message: "AmtNotDetected"}
@@ -126,6 +127,8 @@ var ActivationFailedGetControlMode = CustomError{Code: 134, Message: "Activation
 var ActivationFailedControlMode = CustomError{Code: 135, Message: "ActivationFailed", Details: "recieved invalid control mode"}
 var DuplicateKey = CustomError{Code: 136, Message: "DuplicateKey", Details: "Key pair already exists"}
 var WiredConfigurationFailed = CustomError{Code: 137, Message: "WiredConfigurationFailed"}
+var UnsupportedAMTVersion = CustomError{Code: 138, Message: "UnsupportedAMTVersion"}
+var LMSConnectionFailed = CustomError{Code: 139, Message: "LMSConnectionFailed", Details: "Failed to connect to LMS. Please install LMS for activation."}
 
 // (150-199) Maintenance Errors
 var SyncClockFailed = CustomError{Code: 150, Message: "SyncClockFailed"}